Update decimal-number-to-base-two-converter.py
I made the first loop in a way that resembled my way of thinking about this converter, which was:
Let's pick some decimal numbers.
0.5 -> this is 1/2
0.625 -> can be seen as 0.5 + 0.125 (or 1/2 + 1/8)
0.65625 -> is basically 0.5 + 0.125 + 0.03125 (or 1/2 + 1/8 + 1/32)
0.3 -> you can see it as 1/4 + 1/32 + 1/64 + 1/512 + 1/1024... (and so on, forever) which is 1/2² + 1/2⁵ + 1/2⁶ + 1/2⁹ + 1/2¹⁰...

In binary, if you have n bits, each bit to the left correspond to 2^something (the number 1001, the first bit to the left represents 2³, the second 2², the third 2¹ and the last one 2⁰, so, 1001 means 1*2³ + 0*2² + 0*2¹ + 1*2⁰ -> 8 + 0 + 0 + 1 = 9)
so. each number to the right, in a decimal scale, should correspond to 2^-something, the first one to the right being 1/2⁻¹, the second one being 1/2⁻², the third 1/2⁻³ and so forth.)

so, 0.5 is 1/2, which is 2⁻¹, which is the first binary digit to the right, which represents 0.1
0.625 is 1/2 + 1/8 which is 2⁻¹ + 2⁻³, which in binary represents 0.101
0.3 is 2⁻² + 2⁻⁵ + 2⁻⁶ + 2⁻⁹ + 2⁻¹⁰ which represents 0.01001100110...

if you notice, each 2^-n represents a number "1" in n steps after the dot. 
So, firstly, I made the loop verify if the number is greater than or equal to (1/2 * multiplier), which starts with 1, making the expression 1/2 * multiplier = 1/2. 
If it is greater than or equal to 1/2, it will ad a "1" in the first  position at the list. then, it will subtract this very 1/2 from the user's input. If, contrarily, the number is lesser than 1/2 * multiplier, it will add a "0" to the list in its first position, the it will divide the multiplier by 2 and restart the loop, meaning:

-if the input was 0.5, it will add a "1" to the list and subtract 0.5 from 0.5, leaving 0 and print only 0's (the output would be: 0.10000)
-if the input was 0.625, it will add "1" to the list, as 0.625>0.5, adding a "1" to the first position of the list, then subtracting 0.5 from 0.625, leaving 0.125, it will then, multiply 1 by 1/2 and continue the loop. now, it will check if 0.125 >= 1/2 * 1/2 (half the new multiplier), which is false, since 0.125 >= 0.25 is false. making it add "0" to the second position of the list. it will then multiply 1/2 by 1/2 leaving 1/4. Now, the loop will check if 0.125 >= 1/4 * 1/2. which is true, since 0.125 = 1/8, so it will add a "1" in the third position, and subtract 0.125 from 0.125, leaving zero, and thus, adding 0's to the list from the position 4 and beyond, making the output (0.101000)